github_automation/requirements.txt
"""Project structure:

GITHUB_AUTOMATION/
│── config.py          # Stores GitHub API token & configurations
│── git_operations.py  # Handles Git commands (branching, committing, merging)
│── github_api.py      # Interacts with GitHub API (creating repos, PRs)
│── main.py            # CLI interface for GitHub automation
│── requirements.txt   # Dependencies for the project

"""

"""
1. Clone the Repo
git clone https://github.com/your-username/GITHUB_AUTOMATION.git
cd GITHUB_AUTOMATION


2. Install the dependencies

DEPENDENCIES:

HOW TO INSTALL? ->pip install click(or anyother)

click           # CLI command handling
gitpython       # Interacting with Git repositories
requests        # Making API calls to GitHub


3.Set Up GitHub API Token
Generate a GitHub Personal Access Token (Guide).
Add it to config.py:

GITHUB_TOKEN = "your_github_personal_access_token"


Ensure the token has repo & workflow permissions.

"""